Fix `xml_fatal_err_msg_str` to macro.
I have not fix `%%` to `%` yet.
If this is fixed, `$crate::XML_GET_VAR_STR!` assumes they are replacement
characters. In this case, the number of the arguments for `snprintf` is
not sufficient and it causes an undefined behavior.
This should be fixed after all `$crate::XML_GET_VAR_STR!` are removed.